requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
distro==1.9.0
ecdsa==0.19.2
faiss-cpu==1.13.2
fastapi==0.135.2
filelock==3.25.2
fsspec==2026.2.0
groq==1.1.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==2.30.0
packaging==26.0
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.9.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlitecloud==0.0.84
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
pytest==8.3.4
pytest-asyncio==0.24.0
